NN start
Former-commit-id: 447487def9631bffaaf3f8d43e5866026a0a8cbd

+++ Chess/NonGraphicalBoard.h
@@ -31,7 +31,7 @@ public:
 
     ~NonGraphicalBoard();
 
-    int* Status(bool isWhite);
+    float* Status(bool isWhite);
 
     void SetPiecesAsDefault(Piece** pieces);
